Now put_integer_64_constant takes an INTEGER_64 value.
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@28665 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/Eiffel/eiffel/com_il_generation/Core/core.cs
@@ -540,7 +540,7 @@ public class Core : ICore {
 		core.PutInteger32Constant (i);
 	}
 
-	public void PutInteger64Constant (int i ) {
+	public void PutInteger64Constant (long i ) {
 		core.PutInteger64Constant (i);
 	}
